+++ CHANGELOG.md
@@ -52,6 +52,7 @@ Removed the following long deprecated elements:
 ### Changed
 
 - [package]: Updated dependencies.
+- [platform]: Bumped MatterbridgePlatform v.1.4.0.
 - [deviceManager]: Bumped DeviceManager v.1.1.1.
 - [pluginManager]: Bumped PluginManager v.1.3.1.
 - [broadcastServer]: Bumped BroadcastServer v.1.0.3.